requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
PyJWT==2.10.1
sqlparse==0.5.4
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0




